requirements/base.txt
Django~=2.1.8
PyMySQL
djangorestframework
drf-writable-nested
django-filter
django-mysql
django-environ
